:sparkles: refactor package
